+++ releasenotes.md
@@ -12,6 +12,18 @@ Release Notes
   * `Dlocal`
   * `Onpay`
   * `Safecharge`
+* Documented the new "test" disposition.
+* Added support for the `/disposition/risk_label` output in Score, Insights and
+  Factors. This is the available as the `RiskLabel` property of
+  `MaxMind.MinFraud.Response.Disposition`, and is the label of the custom rule
+  that was triggered by the transaction.
+* Added support for the `/credit_card/was_3d_secure_successful` input in Score,
+  Insights and Factors. This input should indicate whether or not the outcome
+  of 3-D Secure verification (e.g. Safekey, SecureCode, Verified by Visa) was
+  successful. `true` if customer verification was successful, or `false` if
+  the customer failed verification. If 3-D Secure verification was not used, was
+  unavailable, or resulted in another outcome other than success or failure, do
+  not include this field.
 
 3.1.0 (2021-02-02)
 ------------------